Tests/LibWeb/Text/expected/Editing/execCommand-createLink.txt

d1 contents: "foo<a href="https://ladybird.org">bar</a>"
d2 contents: "<a href="https://ladybird.org">ladybird</a>"
